Refactor useable classes into source
