+++ sahha-android/build.gradle
@@ -9,7 +9,7 @@ def secretsPropertiesFile = rootProject.file("sahha-android/secrets.properties")
 def secretsProperties = new Properties()
 secretsProperties.load(new FileInputStream(secretsPropertiesFile))
 
-def gVersionName = "0.12.7"
+def gVersionName = "0.12.8"
 
 android {
     compileSdk 31
