+++ sahha-android/src/main/java/sdk/sahha/android/data/local/AppDatabase.kt
@@ -11,7 +11,7 @@ import sdk.sahha.android.data.remote.dto.SleepDto
 import sdk.sahha.android.domain.model.activities.PreviousActivity
 import sdk.sahha.android.domain.model.activities.RecognisedActivity
 import sdk.sahha.android.domain.model.config.SahhaConfiguration
-import sdk.sahha.android.domain.model.config.SahhaNotificationConfiguration
+import sdk.sahha.android.source.SahhaNotificationConfiguration
 import sdk.sahha.android.domain.model.device.AppUsage
 import sdk.sahha.android.domain.model.device.DeviceUsage
 import sdk.sahha.android.domain.model.device.PhoneUsage
