+++ sahha-android/src/main/java/sdk/sahha/android/data/repository/RemoteRepoImpl.kt
@@ -11,8 +11,6 @@ import sdk.sahha.android.common.*
 import sdk.sahha.android.common.security.Decryptor
 import sdk.sahha.android.common.security.Encryptor
 import sdk.sahha.android.data.Constants.MAX_STEP_POST_VALUE
-import sdk.sahha.android.data.Constants.STEP_COUNTER_DATA_SOURCE
-import sdk.sahha.android.data.Constants.STEP_DETECTOR_DATA_SOURCE
 import sdk.sahha.android.data.Constants.UERT
 import sdk.sahha.android.data.Constants.UET
 import sdk.sahha.android.data.local.dao.DeviceUsageDao
@@ -27,6 +25,7 @@ import sdk.sahha.android.domain.model.auth.TokenData
 import sdk.sahha.android.domain.model.steps.StepData
 import sdk.sahha.android.domain.repository.RemoteRepo
 import sdk.sahha.android.source.Sahha
+import sdk.sahha.android.source.SahhaConverterUtility
 import sdk.sahha.android.source.SahhaDemographic
 import sdk.sahha.android.source.SahhaSensor
